% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntStats.R
\name{calcIntStats}
\alias{calcIntStats}
\title{Calculate mean, var, size factor, phi, & shrunken phi}
\usage{
calcIntStats(counts)
}
\arguments{
\item{counts}{\link[mcCNV:validObjects]{counts object}}
}
\description{
Calculate mean, var, size factor, phi, & shrunken phi
}
