% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareReal2Sim.R
\name{compareReal2Sim}
\alias{compareReal2Sim}
\title{Calculate interval statistics for real & simulated pool based on real}
\usage{
compareReal2Sim(p, seed = 1234)
}
\arguments{
\item{p}{character of length 1, pool name}

\item{seed}{seed value passed to cnvSimPool}
}
\description{
Calculate interval statistics for real & simulated pool based
on the interval created from the real counts
}
