% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPred.R
\name{evalPred}
\alias{evalPred}
\title{Evaluate predictions}
\usage{
evalPred(x, y)
}
\arguments{
\item{x}{logical, prediction of TRUE (positive) or FALSE (negative)}

\item{y}{logical, the "truth" set to compare x to}
}
\description{
Evaluate model predictions
}
