% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLogNorm.R
\name{fitLogNorm}
\alias{fitLogNorm}
\title{Calculate log-normal fit of capture probabilities}
\usage{
fitLogNorm(p, plt = TRUE)
}
\arguments{
\item{p}{character of length 1, pool name}

\item{plt}{logical of length 1, draw the histogram w/ fit when TRUE}
}
\description{
Calculate log-normal fit of capture probabilities
}
