% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltBestPrior}
\alias{pltBestPrior}
\title{Plot best prior across sequencing depth & variant width}
\usage{
pltBestPrior(dat)
}
\arguments{
\item{dat}{data.table w/ prior, factor giving sequencing depth (dep), and
factor giving variant width (fwid)}
}
\description{
Plot best prior across sequencing depth & variant width
}
