% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltECDF}
\alias{pltECDF}
\title{Plot ECDF curves comparing two vectors}
\usage{
pltECDF(v1, v2, q = 0.99, xlab)
}
\arguments{
\item{v1}{vector 1 values, gray solid line}

\item{v2}{vector 2 values, black dotted line}
}
\description{
Plot ECDF curves comparing two vectors
}
