% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltFalseCalls}
\alias{pltFalseCalls}
\title{Plot number of false calls by depth}
\usage{
pltFalseCalls(res)
}
\arguments{
\item{res}{data.table w/ 'ACT', 'anyCNV', 'PRO', 'dep', & 'N' columns; e.g.
\link{sRes$noVariants$mcCNV}}
}
\description{
Plot number of false calls by depth
}
