% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltMnVrCont}
\alias{pltMnVrCont}
\title{Plot variance by mean contours}
\usage{
pltMnVrCont(
  dat,
  grpVec,
  colVec = palette()[unique(grpVec)],
  nbin = 100,
  ncontour = 20,
  lgnd = TRUE
)
}
\arguments{
\item{dat}{subset of \link{realMeanVar}}

\item{grpVec}{factor, same length as \code{nrow(dat)}; defines groups for
plotting}

\item{colVec}{vector of colors, same length as \code{levels(grpVec)}}

\item{nbin}{number of bins in each dimension to aggregate data}

\item{ncontour}{number of contours to plot}
}
\description{
Plot variance by mean contours
}
