% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltPriorMatrix}
\alias{pltPriorMatrix}
\title{Plot prior by depth performance matrix}
\usage{
pltPriorMatrix(mnDat, wid = 1, stat = "mcc")
}
\arguments{
\item{wid}{variant width to plot}

\item{stat}{character, the statistic from \link{procRes} to plot}

\item{mndat}{object$mnDat, where object is output from \link{procRes}}
}
\description{
Plot prior by depth performance matrix
}
