% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltStat}
\alias{pltStat}
\title{Plot procRes results}
\usage{
pltStat(res, stat, ylab = toupper(stat), h = NULL)
}
\arguments{
\item{res}{simulation result summary, output from \link{procRes}}

\item{stat}{character, the statistic from \link{procRes} to plot}

\item{ylab}{character, text for ylab}

\item{h}{numeric, value to draw dashed horizontal line}
}
\description{
Plot procRes results
}
