% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltStatCompare}
\alias{pltStatCompare}
\title{Plot summary statistics comparing two fits}
\usage{
pltStatCompare(xRes, yRes, stat, xlab = "X", ylab = "Y", lbl = toupper(stat))
}
\arguments{
\item{stat}{character, the statistic from \link{procRes} to plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{lbl}{character, the label for the upper left corner}

\item{xMn}{x-axis data; object$mnDat, where object is output from \link{procRes}}

\item{yMn}{y-axis data; object$mnDat, where object is output from \link{procRes}}
}
\description{
Plot summary statistics comparing two fits
}
