% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltSubjectStatByPool}
\alias{pltSubjectStatByPool}
\title{Plot subject stat by pool}
\usage{
pltSubjectStatByPool(stat, poolVec = NULL, logY = TRUE, ylab = stat)
}
\arguments{
\item{stat}{character giving the stat to plot, e.g. totalMolCount}

\item{poolVec}{character vector giving the pools to plot}

\item{logY}{logical, use log y-axis when TRUE}

\item{ylab}{(optional) alternate y-axis label; defaults to stat}
}
\description{
Plot subject stat, e.g. totalMolCount, by pool
}
