% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procRes.R
\name{procRes}
\alias{procRes}
\title{Process call summaries}
\usage{
procRes(x)
}
\arguments{
\item{x}{results of simulation study}
}
\description{
Takes simulation study results, and calculates confusion matrix
statistics based.
}
