% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCounts.R
\name{subsetCounts}
\alias{subsetCounts}
\title{Make counts object from subjectCounts}
\usage{
subsetCounts(subjects)
}
\arguments{
\item{subjects}{vector of subjects from \link{subjectCounts}}
}
\value{
mcCNV counts object for the given subjects
}
\description{
Make \link[mcCNV:validObjects]{counts object} from \link{subjectCounts}
}
