% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pltAlpha0}
\alias{pltAlpha0}
\title{Plot alpha0 estimates by pool}
\usage{
pltAlpha0(a0tbl)
}
\arguments{
\item{a0tbl}{data.table containing alpha0 values and other summary info}
}
\description{
Plot alpha0 estimates by pool
}
